spec file works
Former-commit-id: 05be0493eb59f5a356b0fd18239cd0b41e445000

+++ admin/svn2tgz.sh
@@ -1,19 +1,35 @@
 #!/bin/bash
 
-# Makes the current m8r development version into a tarball, removing svn dirs
+# Makes the current m8r development version into a tarball, without svn dirs
 # Useful for installing on machines that do not have svn.
- 
-# This is also a prerequisite for making RPMs. Those would be made from a future
-# stable release, but for now we need to practice using the development version
 
-SRC=rsfsrc
+# Argument 1: the directory and tarball name. If not given, default is RSFSRC
+# Argument 2: SVN version number. If not given, default is current version
+# Argument 3: Whether to rm -rf ('clean') or not (no arg given) the downloaded dir
+
+
+# Argument 1 -- tarball and directory name:
+if [ -z $1 ]; then
+  SRC='RSFSRC'
+else
+  SRC=$1
+fi
+
+# Argument 2 -- SVN version number:
+if [ -z $2 ]; then
+  REL='-r '$2
+else
+  REL=''
+fi
 
 # Get a fresh copy from the repository, guaranteed not to have local junk in it
-svn co https://rsf.svn.sourceforge.net/svnroot/rsf/trunk $SRC
+svn export $REL https://rsf.svn.sourceforge.net/svnroot/rsf/trunk $SRC
 
-rm -rf $SRC/{admin,user/nobody,api/octave}
-find $SRC -name '.svn' -type d -exec rm -rf {} \;
+rm -rf $SRC/{admin,user/nobody,api/octave,scons}
 
-tar -cvzf $SRC.tgz $SRC
+tar -cvzf $SRC.tar.gz $SRC
 
-rm -rf $SRC
+# Argument 3 -- cleanup:
+if test "$3" = "clean"; then
+  rm -rf $SRC
+fi
